% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance_history}
\alias{balance_history}
\title{Get the balance history of your Stripe account}
\usage{
balance_history()
}
\value{
List object
}
\description{
Get the balance history of your Stripe account
}

