% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance_transaction}
\alias{balance_transaction}
\title{Get the balance transaction}
\usage{
balance_transaction(transactionId)
}
\arguments{
\item{transactionId}{The ID of the balance transaction

Balance transaction ID found on any API object that affects
  balance e.g. a charge or transfer}
}
\value{
List object
}
\description{
Get the balance transaction
}
