% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{cancel_subscription}
\alias{cancel_subscription}
\title{Cancel a subscription}
\usage{
cancel_subscription(customerId, subscriptionId, at_period_end = FALSE)
}
\arguments{
\item{customerId}{The ID of the customer}

\item{subscriptionId}{The ID of the subscription}

\item{at_period_end}{Default FALSE. Delay cancellation to end of billing period.}
}
\value{
List object
}
\description{
Cancel a subscription
}
\seealso{
Other subscriptions: \code{\link{create_subscription}},
  \code{\link{get_subscription}},
  \code{\link{list_subscriptions}},
  \code{\link{update_subscription}}
}
\concept{subscriptions}
