% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/charge.R
\name{capture_charge}
\alias{capture_charge}
\title{Capture a charge}
\usage{
capture_charge(chargeId, amount = NULL, application_fee = NULL,
  receipt_email = NULL, statement_descriptor = NULL)
}
\arguments{
\item{chargeId}{A charge from a previous uncaptured charge}

\item{amount}{Amount in smallest currency unit}

\item{application_fee}{Optional charge fee.}

\item{receipt_email}{email address to send receipt to}

\item{statement_descriptor}{22 chars displayed on customer statement}
}
\value{
A charge object
}
\description{
Capture a charge
}
\seealso{
Other charges: \code{\link{charge_card}};
  \code{\link{get_charge}}; \code{\link{list_charges}};
  \code{\link{update_charge}}
}

