% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{charge_card}
\alias{charge_card}
\title{Charge a credit card}
\usage{
charge_card(amount, currency, customer = NULL, source = NULL,
  receipt_email = NULL, application_fee = NULL, capture = TRUE,
  description = NULL, destination = NULL, metadata = NULL,
  shipping = NULL)
}
\arguments{
\item{amount}{Amount in smallest currency unit}

\item{currency}{ISO code for currency}

\item{customer}{ID of existing customer. This or source required.}

\item{source}{A payment source to be charged}

\item{receipt_email}{email address to send receipt to}

\item{application_fee}{Optional charge fee.}

\item{capture}{Default TRUE}

\item{description}{Default NULL}

\item{destination}{Account to charge from}

\item{metadata}{A named list of metadata}

\item{shipping}{Shipping info for the charge}

\item{statement_descriptor}{22 chars displayed on customer statement}
}
\value{
List object
}
\description{
Charge a credit card
}

