% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/requests.R
\name{checkRequest}
\alias{checkRequest}
\title{Check request content}
\usage{
checkRequest(req)
}
\arguments{
\item{req}{A httr request}
}
\value{
content if it is there, or NULL
}
\description{
Check request content
}
\keyword{internal}

