% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{create_bank_token}
\alias{create_bank_token}
\title{Create a bank account token}
\usage{
create_bank_token(account_number, country, currency,
  routing_number = NULL, name = NULL, account_holder_type = NULL)
}
\arguments{
\item{account_number}{The bank account number}

\item{country}{Country of the bank}

\item{currency}{Currency of the bank}

\item{routing_number}{Or sort code for bank.}

\item{name}{Bank account holder's full name}

\item{account_holder_type}{"individual" or "company"}
}
\description{
Store this token instead of bank details
}
\details{
\code{routing_number} is not required if
  \code{account_number} is an IBAN
}
\seealso{
Other tokens: \code{\link{create_card_token}},
  \code{\link{get_token}}
}
\concept{tokens}
