% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{create_card_token}
\alias{create_card_token}
\title{Create a card token}
\usage{
create_card_token(number, exp_month, exp_year, cvc, name = NULL)
}
\arguments{
\item{number}{The card number}

\item{exp_month}{Expiry month (two digits)}

\item{exp_year}{Expiry year (two or four digits)}

\item{cvc}{Card security code}

\item{name}{Cardholder's full name}
}
\description{
Store this token instead of card details
}
\details{
Not implemented: address
}
\seealso{
Other tokens: \code{\link{create_bank_token}},
  \code{\link{get_token}}
}
