% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{create_customer}
\alias{create_customer}
\title{Create a customer}
\usage{
create_customer(account_balance = NULL, coupon = NULL, description = NULL,
  email = NULL, metadata = NULL, plan = NULL, quantity = NULL,
  shipping = NULL, source = NULL, tax_percent = NULL, trail_end = NULL)
}
\arguments{
\item{account_balance}{Integer amount in pence}

\item{coupon}{A discount on recurring charges}

\item{description}{Arbitary string}

\item{email}{customer's email address}

\item{metadata}{Set of key/value pairs}

\item{plan}{ID of the plan to subscribe customer to}

\item{quantity}{Quantity to apply to subscription. Needs plan.}

\item{shipping}{Optional dictionary}

\item{source}{A token from \link{get_token}}

\item{tax_percent}{A positive decimal. Percentage added as tax.}

\item{trail_end}{Unix timestamp when trial period ends. Needs plan.}
}
\value{
List object
}
\description{
Create a customer
}
\details{
Setting \code{account_balance} to negative means the customer will have a credit.
  A positive amount means it will be added to the next invoice.

  \code{description, email, metadata} will be set to empty if you POST an empty value.
}
\seealso{
Other customers: \code{\link{delete_customer}},
  \code{\link{get_customer}}, \code{\link{list_customers}},
  \code{\link{update_customer}}
}

