% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{create_subscription}
\alias{create_subscription}
\title{Create a subscription for a customer}
\usage{
create_subscription(customerId, plan, coupon = NULL, source = NULL,
  quantity = NULL, metadata = NULL, tax_percent = NULL,
  trial_end = NULL)
}
\arguments{
\item{customerId}{The ID of the customer}

\item{plan}{ID of the plan to subscribe customer to}

\item{coupon}{A discount on recurring charges}

\item{source}{A token from \link{get_token}}

\item{quantity}{Default is 1}

\item{metadata}{A named list}

\item{tax_percent}{Postive decimal}

\item{trial_end}{Unix timestamp or 'now'}
}
\value{
List object
}
\description{
Create a subscription for a customer
}
\details{
Not implemented: application_fee_percent
}
\seealso{
Other subscriptions: \code{\link{cancel_subscription}},
  \code{\link{get_subscription}},
  \code{\link{list_subscriptions}},
  \code{\link{update_subscription}}
}
