% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{delete_customer}
\alias{delete_customer}
\title{Delete a customer}
\usage{
delete_customer(customerId)
}
\arguments{
\item{customerId}{The ID of the customer}
}
\value{
List object
}
\description{
Delete a customer
}
\seealso{
Other customers: \code{\link{create_customer}},
  \code{\link{get_customer}}, \code{\link{list_customers}},
  \code{\link{update_customer}}
}
