% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{get_charge}
\alias{get_charge}
\title{Retieve a charge}
\usage{
get_charge(chargeId)
}
\arguments{
\item{chargeId}{The ID from a previous charge}
}
\value{
List object
}
\description{
Retieve a charge
}
\seealso{
Other charges: \code{\link{capture_charge}},
  \code{\link{charge_card}}, \code{\link{list_charges}},
  \code{\link{update_charge}}
}
\concept{charges}
