% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{get_customer}
\alias{get_customer}
\title{Retrieve a customer}
\usage{
get_customer(customerId)
}
\arguments{
\item{customerId}{The ID of the customer}
}
\value{
List object
}
\description{
Retrieve a customer
}
\seealso{
Other customers: \code{\link{create_customer}},
  \code{\link{delete_customer}},
  \code{\link{list_customers}},
  \code{\link{update_customer}}
}

