% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{get_subscription}
\alias{get_subscription}
\title{Retrieve a subscription}
\usage{
get_subscription(customerId, subscriptionId)
}
\arguments{
\item{customerId}{The ID of the customer}

\item{subscriptionId}{The ID of the subscription}
}
\value{
List object
}
\description{
Retrieve a subscription
}
\seealso{
Other subscriptions: \code{\link{cancel_subscription}},
  \code{\link{create_subscription}},
  \code{\link{list_subscriptions}},
  \code{\link{update_subscription}}
}

