% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_token}
\alias{get_token}
\title{Retrieve a token}
\usage{
get_token(token)
}
\arguments{
\item{token}{The ID of the token}
}
\value{
Token object
}
\description{
A token as created by \link{create_card_token}
}
\seealso{
Other tokens: \code{\link{create_bank_token}},
  \code{\link{create_card_token}}
}
