% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{idempotency}
\alias{idempotency}
\title{Idempotency}
\usage{
idempotency(seed = 1234567)
}
\arguments{
\item{seed}{Random seed}
}
\value{
A random 15 digit hash
}
\description{
A random code to ensure no repeats
}
