% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{list_customers}
\alias{list_customers}
\title{List all customers}
\usage{
list_customers(created = NULL, ending_before = NULL, limit = NULL,
  starting_before = NULL)
}
\arguments{
\item{created}{filter on created}

\item{ending_before}{filter for pagination}

\item{limit}{Between 1 and 100}

\item{starting_after}{filter for pagination}
}
\value{
List object
}
\description{
List all customers
}

