% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_meta}
\alias{make_meta}
\title{Make Metadata keys}
\usage{
make_meta(metadata)
}
\arguments{
\item{meta_list}{A named list of meta data}
}
\value{
meta data suitable to pass to API
}
\description{
Can have 20 keys total, keynames < 40, values < 500
}
\keyword{internal}
