% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{observeCharge}
\alias{observeCharge}
\title{Charge Observer}
\usage{
observeCharge(status, input, amount, currency = "gbp", plan = NULL)
}
\arguments{
\item{status}{A reactive list with element \code{charged}}

\item{input}{Shiny input object}

\item{amount}{The amount to charge}

\item{currency}{Currency of amount}

\item{plan}{Optional payment plan to subscribe to.}
}
\description{
Needs to be present to do the charge with stripeForm
}
\seealso{
Other shiny: \code{\link{renderStripeForm}},
  \code{\link{stripeFormOutput}},
  \code{\link{stripeShinyInit}}
}

