% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderStripeForm}
\alias{renderStripeForm}
\title{Render Shiny payment form}
\usage{
renderStripeForm(status, amount, plan = "",
  bottom_left = "Powered by Stripe", thanks = p(amount,
  " paid successful.  Thanks!"))
}
\arguments{
\item{status}{A reactive list with element \code{charged}}

\item{amount}{String with amount to charge e.g. "€4.40"}

\item{plan}{Optional payment plan to show}

\item{bottom_left}{Text shown at bottom left}

\item{thanks}{The thank you UI upon payment}
}
\value{
A StripeR form to output with \link{stripeFormOutput} The form has
  these inputs: input$email input$card_number input$exp_month input$exp_year
  input$cvc

  Action button: input$charge_card
}
\description{
Render Shiny payment form
}
\seealso{
Other shiny: \code{\link{observeStripeCharge}},
  \code{\link{stripeFormOutput}},
  \code{\link{stripeShinyInit}}
}

