% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{retrieve_charge}
\alias{retrieve_charge}
\title{Retieve a charge}
\usage{
retrieve_charge(chargeId)
}
\arguments{
\item{chargeId}{The ID from a previous charge}
}
\value{
List object
}
\description{
Retieve a charge
}

