% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{retrieve_customer}
\alias{retrieve_customer}
\title{Retrieve a customer}
\usage{
retrieve_customer(customerId)

retrieve_customer(customerId)
}
\arguments{
\item{customerId}{The ID of the customer}

\item{customerId}{The ID of the customer}
}
\value{
List object

List object
}
\description{
Retrieve a customer

Delete a customer
}

