% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{stripeFormOutput}
\alias{stripeFormOutput}
\title{stripeForm Shiny Output}
\usage{
stripeFormOutput(outputId)
}
\arguments{
\item{outputId}{What \link{renderStripeForm} outputs to.}
}
\value{
The Stripe Form in Shiny UI.
}
\description{
stripeForm Shiny Output
}
\seealso{
Other shiny: \code{\link{renderStripeForm}}
}

