% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\docType{data}
\name{stripeR_auth}
\alias{stripeR_auth}
\title{R6 environment to store authentication credentials}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stripeR_auth
}
\description{
Used to keep persistent state.
}
\keyword{datasets}

