% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{stripeR_init}
\alias{stripeR_init}
\title{Set global configurations}
\usage{
stripeR_init(live = FALSE)
}
\arguments{
\item{live}{TRUE If the payments should be against your live Stripe account.}
}
\description{
Run this at the start of every stripeR session.
}
\details{
Set up your Stripe API keys first via:

\code{options('stripeR.secret_test') <- SECRET_TEST_KEY}
\code{options('stripeR.secret_live') <- SECRET_LIVE_KEY}
\code{options('stripeR.public_test') <- PUBLIC_TEST_KEY},
\code{options('stripeR.public_live') <- PUBLIC_LIVE_KEY}

Then run \code{stripeR_init()} to initialise your Stripe API session.
}

