% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{stripeRShinyUI}
\alias{stripeRShinyUI}
\title{stripeRShiny UI}
\usage{
stripeRShinyUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
A Shiny Stripe From
}
\description{
Shiny Module for use with \link{stripeRShiny}
}
\examples{

\dontrun{
## server.R
library(shiny)
library(stripeR)
function(input, output, session){

callModule(stripeRShiny, "stripe1",
           amount=2000,
           plan="example",
           formAmount=reactive("$20.00"),
           formText=reactive("Please pay $20.00"))
           }

## ui.R
library(shiny)
library(stripeR)
fluidPage(
titlePanel("StripeR Demo"),
# A Stripe Form
sidebarLayout(
  sidebarPanel(
    stripeRShinyUI("stripe1")
 ),

mainPanel(...)))

}
}
