% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{stripeShinyInit}
\alias{stripeShinyInit}
\title{Stripe Shiny status object}
\usage{
stripeShinyInit()
}
\description{
Use at start of Shiny session
}
\seealso{
Other shiny: \code{\link{observeCharge}},
  \code{\link{renderStripeForm}},
  \code{\link{stripeFormOutput}}
}

