% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{update_charge}
\alias{update_charge}
\title{Update a charge}
\usage{
update_charge(chargeId, description = NULL, fraud_details = NULL,
  metadata = NULL, receipt_email = NULL, shipping = NULL)
}
\arguments{
\item{chargeId}{Id of charge}

\item{description}{Default NULL}

\item{fraud_details}{Possible fraud information}

\item{metadata}{A named list of metadata}

\item{receipt_email}{email address to send receipt to}

\item{shipping}{Shipping info for the charge}
}
\value{
List object
}
\description{
Update a charge
}
\seealso{
Other charges: \code{\link{capture_charge}},
  \code{\link{charge_card}}, \code{\link{get_charge}},
  \code{\link{list_charges}}
}

