% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{update_subscription}
\alias{update_subscription}
\title{Update a subscription for a customer}
\usage{
update_subscription(subscriptionId, customerId, plan, prorate = TRUE,
  proration_date = NULL, coupon = NULL, source = NULL, quantity = NULL,
  metadata = NULL, tax_percent = NULL, trial_end = NULL)
}
\arguments{
\item{subscriptionId}{The ID of the subscription}

\item{customerId}{The ID of the customer}

\item{plan}{ID of the plan to subscribe customer to}

\item{prorate}{Whether to prorate switching plans during billing cycle}

\item{proration_date}{Timestamp of proration effective. Default current time}

\item{coupon}{A discount on recurring charges}

\item{source}{A token from \link{get_token}}

\item{quantity}{Default is 1}

\item{metadata}{A named list}

\item{tax_percent}{Postive decimal}

\item{trial_end}{Unix timestamp or 'now'}
}
\value{
List object
}
\description{
Update a subscription for a customer
}
\details{
Not implemented: application_fee_percent
}
\seealso{
Other subscriptions: \code{\link{cancel_subscription}},
  \code{\link{create_subscription}},
  \code{\link{get_subscription}},
  \code{\link{list_subscriptions}}
}

