% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{updateStatus}
\alias{updateStatus}
\title{updateStatus}
\usage{
updateStatus(attempt, status)
}
\arguments{
\item{attempt}{stripeR request}

\item{status}{status object ot update}
}
\value{
Updates the status object with values in attempt
}
\description{
updateStatus
}
\seealso{
Other shiny: \code{\link{stripeShinyInit}}
}
\concept{shiny}
\keyword{internal}
