% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{charge_card}
\alias{charge_card}
\title{Charge a credit card}
\usage{
charge_card(amount, currency, customer = NULL, source = NULL,
  receipt_email = NULL, application_fee = NULL, capture = TRUE,
  description = NULL, metadata = NULL, shipping = NULL,
  statement_descriptor = NULL)
}
\arguments{
\item{amount}{Amount in smallest currency unit}

\item{currency}{ISO code for currency}

\item{customer}{ID of existing customer. This or source required.}

\item{source}{A payment source to be charged}

\item{receipt_email}{email address to send receipt to}

\item{application_fee}{Optional charge fee.}

\item{capture}{TRUE to charge now, FALSE charge with \link{capture_charge}}

\item{description}{Default NULL}

\item{metadata}{A named list}

\item{shipping}{Shipping info for the charge}

\item{statement_descriptor}{22 chars displayed on customer statement}
}
\value{
List object
}
\description{
Charge a credit card
}
\details{
\code{source} and \code{customer}:
If you pass a customer ID, the source must be the ID of
  a source belonging to the customer.

If you do not pass a customer ID,
  the source you provide must be a token from \link{create_token}
}
\seealso{
Other charges: \code{\link{capture_charge}},
  \code{\link{get_charge}}, \code{\link{list_charges}},
  \code{\link{update_charge}}
}

