% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/requests.R
\name{do_request}
\alias{do_request}
\title{Do a request}
\usage{
do_request(url, request_type, the_body = NULL, customConfig = NULL)
}
\arguments{
\item{url}{The url of the request}

\item{request_type}{GET, POST, PUT etc.}

\item{the_body}{Body to send with the request, if any}

\item{customConfig}{a list of custom configurations from httr}
}
\value{
the request content or NULL
}
\description{
Do a request
}
\keyword{internal}

