% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderStripeForm}
\alias{renderStripeForm}
\title{Render Shiny payment form}
\usage{
renderStripeForm(amount)
}
\arguments{
\item{amount}{Amount to charge}
}
\value{
A StripeR form to output with \link{stripeFormOutput}
}
\description{
Render Shiny payment form
}
\seealso{
Other shiny: \code{\link{stripeFormOutput}}
}

