% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderStripeForm}
\alias{renderStripeForm}
\title{Render Shiny payment form}
\usage{
renderStripeForm(status, bottom_left = "Powered by Stripe",
  thanks = p("Thanks!"))
}
\arguments{
\item{status}{A reactive list with element \code{charged}}

\item{bottom_left}{Text shown at bottom left}

\item{thanks}{The thank you UI upon payment}
}
\value{
A StripeR form to output with \link{stripeFormOutput} The form has
  these inputs: input$email input$card_number input$exp_month input$exp_year
  input$cvc

  Action button: input$charge_card
}
\description{
Render Shiny payment form
}
\details{
You need to set the formAmount and payment plan
    via status$formAmount and status$formText
}
\seealso{
Other shiny: \code{\link{observeStripeCharge}},
  \code{\link{stripeFormOutput}},
  \code{\link{stripeShinyInit}}
}

