% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{stripeRShiny}
\alias{stripeRShiny}
\title{stripeRShiny}
\usage{
stripeRShiny(input, output, session, amount, plan, formAmount, formText,
  bottom_left = "Powered by StripeR", thanks = shiny::p("Thanks!"),
  currency = "gbp", live = FALSE, metadata = NULL)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{amount}{Amount to charge, reactive}

\item{plan}{reactive - If to place the customer on a plan configured in Stripe}

\item{formAmount}{reactive - Text to display for amount in form}

\item{formText}{reactive - Custom text for form}

\item{bottom_left}{What to see in the bottom left of form}

\item{thanks}{What to see once the form is successfully submitted}

\item{live}{Whether to charge against the live Stripe account.}

\item{metadata}{reactive - A named list of other data to send into stripeR}
}
\value{
A reactive status object
}
\description{
Shiny Module for use with \link{stripeRShinyUI}
}
\details{
Call via status <- shiny::callModule(stripeRShiny, "your_id")
}

