% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{stripeRShinyUI}
\alias{stripeRShinyUI}
\title{stripeRShiny UI}
\usage{
stripeRShinyUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
A Shiny Stripe From
}
\description{
Shiny Module for use with \link{stripeRShiny}
}
