% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{stripeShinyInit}
\alias{stripeShinyInit}
\title{Stripe Shiny status object}
\usage{
stripeShinyInit(...)
}
\arguments{
\item{...}{Reactive data you can send into Stripe metadaata}
}
\description{
Use at start of Shiny session
}
\details{
If a user has paid the $charged = TRUE else FALSE.

  e.g. if status <- stripeShinyInit()
  if(status$charged) "Thanks for paying"
}
\seealso{
Other shiny: \code{\link{observeStripeCharge}},
  \code{\link{renderStripeForm}},
  \code{\link{stripeFormOutput}}
}

